Move follow routes to krate::follow::*
